requirements.txt
Django
djangorestframework
psycopg2
Pillow

flake8
